Ensure ingested data has a value for `last_modified`.
As an `auto_now` field cannot be null, but `loaddata` (which is used by this command) bypasses both `Model.save()` and 'QuerySet.update()` by going directly to the database, data ingestion causes an IntegrityError unless the `last_modified` field's value is explicitly set. The value used is `datetime.now()` to replicate the normal behaviour when an object with an `auto_now` field is created.