Require Hawk authentication for Activity Stream feed.
The content type hack in `authorisation.py` is necessary to get this working with Django's WSGI implementation, as it adds a `text/plain` value even though AS sent a request without any value. Strictly speaking, it isn't necessary if (as I assume is the case) Gunicorn doesn't make the same mistake of setting a default value for the request header. But in that case, it won't do any harm, and this at least lets one test locally.